chore: move validateBurnedAmount into its own folder
