+++ src/utils/index.ts
@@ -13,6 +13,6 @@ export * from './getTransferableInputsByTx';
 export * from './getTransferableOutputsByTx';
 export * from './getUtxoInfo';
 export * from './getBurnedAmountByTx';
-export * from './validateBurnedAmount';
+export * from './validateBurnedAmount/validateBurnedAmount';
 export * from './isEtnaEnabled';
 export { unpackWithManager, getManagerForVM, packTx } from './packTx';
