+++ src/vms/pvm/index.ts
@@ -1,6 +1,7 @@
 export * from './builder';
 export * from './models';
 export * from './api';
+export { calculateFee } from './txs/fee/calculator';
 
 // Exposed Etna builder functions under `e` namespace
 export * as e from './etna-builder';
